test_api/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
contextlib2==21.6.0
Django==5.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-ninja==1.3.0
django-ninja-extra==0.21.4
exceptiongroup==1.2.2
fastapi==0.112.1
idna==3.7
injector==0.22.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.2.2
